+++ flagscale/serve/utils.py
@@ -23,27 +23,6 @@ def load_once(func):
     return wrapper
 
 
-def flatten_dict_to_args(config_dict, ignore_keys=[]):
-    args = []
-    for key, value in config_dict.items():
-        if key in ignore_keys:
-            continue
-        key = key.replace("_", "-")
-        if isinstance(value, dict):
-            args.extend(flatten_dict_to_args(value, ignore_keys))
-        elif isinstance(value, list):
-            args.append(f"--{key}")
-            for v in value:
-                args.append(f"{v}")
-        elif isinstance(value, bool):
-            if value:
-                args.append(f"--{key}")
-        else:
-            args.append(f"--{key}")
-            args.append(f"{value}")
-    return args
-
-
 @load_once
 def load_args() -> None:
     """Load configuration for cluster init"""
