+++ flagscale/utils.py
@@ -55,3 +55,24 @@ class CustomModuleLoader:
 
         sys.meta_path.insert(0, finder)
         return module
+
+
+def flatten_dict_to_args(config_dict, ignore_keys=[]):
+    args = []
+    for key, value in config_dict.items():
+        if key in ignore_keys:
+            continue
+        key = key.replace("_", "-")
+        if isinstance(value, dict):
+            args.extend(flatten_dict_to_args(value, ignore_keys))
+        elif isinstance(value, list):
+            args.append(f"--{key}")
+            for v in value:
+                args.append(f"{v}")
+        elif isinstance(value, bool):
+            if value:
+                args.append(f"--{key}")
+        else:
+            args.append(f"--{key}")
+            args.append(f"{value}")
+    return args
